core: move npm tests to core package and add CI test script
Move npm and npm-config tests from opencode package to core package where

the corresponding source modules live. This centralizes the npm-related

testing logic with the implementation.

Add test:ci script to core package for running tests with JUnit output

format for CI environments.
